Added a check for flag 'continueAfterAbort' to weather database creation
- do not delete existing weather database if flags ‘continueAfterAbort’
and ‘createAndPopulateWeatherDatabase’ are set to TRUE

+++ 2_SWSF_p2of4_CreateDB_Tables_v51.R
@@ -13,8 +13,12 @@ suppressMessages(library(RSQLite))
 #------------------------CREATE WEATHER DATABASE AND POPULATION WITH DAILY WEATHER FOR CURRENT CONDITIONS
 if (createAndPopulateWeatherDatabase) {
 	if (file.exists(dbWeatherDataFile)) {
-		print("Removing old database")
-		file.remove(dbWeatherDataFile)
+		if (continueAfterAbort) {
+			stop("Weather database exists, 'continueAfterAbort' is TRUE, and 'createAndPopulateWeatherDatabase' is TRUE: a maximum of two of these three conditions may simultaneously be TRUE: adjust inputs and restart")
+		} else {
+			print("Removing old database")
+			file.remove(dbWeatherDataFile)
+		}
 	}
 	
 	print("Creating new Weather database")
